+++ server/cpho/jinja2/indicator_data/manage_indicator_data.jinja2
@@ -210,12 +210,9 @@
       let formTemplateNode = document.querySelector(".age-group-empty-form-container")
       let container = document.querySelector(".age-group-form-list")
       let totalFormsInput = document.querySelector("#id_agegroup-TOTAL_FORMS") //management form inputs
-
       e.preventDefault()
-
       let formRegex = RegExp(`agegroup-__prefix__-`,'g')
       let newFormHtml = formTemplateNode.innerHTML.replace(formRegex, `agegroup-${newFormIndex}-`)
-
       let newForm = document.createElement("tr")
       newForm.classList.add("age-group-form")
       newForm.innerHTML = newFormHtml;
